Move LatexFieldFormatter to the bibtex package
