+++ src/main/java/net/sf/jabref/bibtex/BibEntryWriter.java
@@ -10,7 +10,6 @@ import java.util.TreeSet;
 import java.util.function.Predicate;
 
 import net.sf.jabref.Globals;
-import net.sf.jabref.exporter.LatexFieldFormatter;
 import net.sf.jabref.logic.TypedBibEntry;
 import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.EntryTypes;
