+++ src/main/java/net/sf/jabref/exporter/BibDatabaseWriter.java
@@ -36,6 +36,7 @@ import net.sf.jabref.BibDatabaseContext;
 import net.sf.jabref.Globals;
 import net.sf.jabref.MetaData;
 import net.sf.jabref.bibtex.BibEntryWriter;
+import net.sf.jabref.bibtex.LatexFieldFormatter;
 import net.sf.jabref.bibtex.comparator.BibtexStringComparator;
 import net.sf.jabref.bibtex.comparator.CrossRefEntryComparator;
 import net.sf.jabref.bibtex.comparator.FieldComparator;
