+++ src/main/java/net/sf/jabref/gui/StringDialog.java
@@ -47,8 +47,8 @@ import javax.swing.undo.CompoundEdit;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
+import net.sf.jabref.bibtex.LatexFieldFormatter;
 import net.sf.jabref.bibtex.comparator.BibtexStringComparator;
-import net.sf.jabref.exporter.LatexFieldFormatter;
 import net.sf.jabref.gui.actions.Actions;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.gui.help.HelpFiles;
