+++ src/main/java/net/sf/jabref/gui/TransferableBibtexEntry.java
@@ -25,7 +25,7 @@ import java.util.List;
 import javax.swing.JOptionPane;
 
 import net.sf.jabref.bibtex.BibEntryWriter;
-import net.sf.jabref.exporter.LatexFieldFormatter;
+import net.sf.jabref.bibtex.LatexFieldFormatter;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.model.entry.BibEntry;
