+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditor.java
@@ -70,8 +70,8 @@ import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.bibtex.BibEntryWriter;
 import net.sf.jabref.bibtex.FieldProperties;
 import net.sf.jabref.bibtex.InternalBibtexFields;
+import net.sf.jabref.bibtex.LatexFieldFormatter;
 import net.sf.jabref.event.FieldChangedEvent;
-import net.sf.jabref.exporter.LatexFieldFormatter;
 import net.sf.jabref.external.WriteXMPEntryEditorAction;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.EntryContainer;
