+++ src/test/java/net/sf/jabref/logic/xmp/XMPUtilTest.java
@@ -29,7 +29,7 @@ import javax.xml.transform.TransformerException;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.bibtex.BibEntryWriter;
-import net.sf.jabref.exporter.LatexFieldFormatter;
+import net.sf.jabref.bibtex.LatexFieldFormatter;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.importer.fileformat.BibtexParser;
 import net.sf.jabref.model.database.BibDatabaseMode;
